// src/app/onboarding/types/formData.ts
export interface FormData {
    firstName: string;
    lastName: string;
    email: string;
    profileComplete: boolean;
    avatarName: string;
    businessName: string;
    websiteURL: string;
    companySize: string;
    businessIndustry: string;
    customBusinessIndustry: string;
    brandIdentity: string;
    brandGuidelines: string;
    accountType: string;
    userUid: string;
    targetAudience: string[];
    contentGoals: string[];
    contentPreferences: string[];
    creatorType: string;
}